Cancelling. (#26)
Addresses https://github.com/niksy/throttle-debounce/issues/6.

I started this a few days ago but I see there's already kind of an implementation in the time it took me to create this one. I figured I might as well open the PR for at least some discussion. 

I think this impl may be a little more thorough since after cancelling the the callback function will not be called again regardless of whether or not it was going to be executed in a timeout or synchronously. 

This also moves the documentation to the Usage section and provides an example.